+++ app/models/base_model.rb
@@ -1,0 +1,3 @@
+class BaseModel
+  include BaseModelInstanceMethods
+end
